DAOS-16908 control: add flag for TCP firewall feature in control plane
backport #16292.

* Add flag for TCP firewall feature in control plane.

This flag will be set in the daos_server.yml and sent to the client via
the
daos_agent. If the flag is true, we will set a client-side environment
variable
that will be used by Mercury and Libfabric to set a bit in subsequent
RPC requests
to the server. If the flag is false, it will not be set.

Signed-off-by: Jinshan Xiong <jinshanx@google.com>
Signed-off-by: Yokesh Jayakumar <karthj@google.com>